Move test to integration tests
