+++ test/unit/test_options.py
@@ -17,14 +17,11 @@ from dataclasses import asdict
 from ddt import data, ddt
 from pydantic import ValidationError
 from qiskit.transpiler import CouplingMap
-from qiskit_aer.noise import NoiseModel
 
-from qiskit_ibm_runtime.fake_provider import FakeManilaV2, FakeNairobiV2
 from qiskit_ibm_runtime.options import EstimatorOptions, SamplerOptions
 from qiskit_ibm_runtime.runtime_options import RuntimeOptions
 
 from ..ibm_test_case import IBMTestCase
-from ..utils import combine
 
 
 @ddt
@@ -74,29 +71,3 @@ class TestOptionsV2(IBMTestCase):
                 inputs = opt_cls._get_program_inputs(asdict(options))["options"]
                 resulting_cmap = inputs["simulator"]["coupling_map"]
                 self.assertEqual(coupling_map, set(map(tuple, resulting_cmap)))
-
-    @combine(
-        opt_cls=[EstimatorOptions, SamplerOptions], fake_backend=[FakeManilaV2(), FakeNairobiV2()]
-    )
-    def test_simulator_set_backend(self, opt_cls, fake_backend):
-        """Test Options.simulator.set_backend method."""
-        options = opt_cls()
-        options.simulator.seed_simulator = 42
-        options.simulator.set_backend(fake_backend)
-
-        noise_model = NoiseModel.from_backend(fake_backend)
-        basis_gates = fake_backend.operation_names
-        coupling_map = fake_backend.coupling_map
-
-        self.assertEqual(options.simulator.coupling_map, coupling_map)
-        self.assertEqual(options.simulator.noise_model, noise_model)
-
-        expected_options = opt_cls()
-        expected_options.simulator = {
-            "noise_model": noise_model,
-            "basis_gates": basis_gates,
-            "coupling_map": coupling_map,
-            "seed_simulator": 42,
-        }
-
-        self.assertDictEqual(asdict(options), asdict(expected_options))
